+++ Projekt zaliczeniowy PAB/src/dataModels/recipeCRUD.ts
@@ -1,40 +1,58 @@
-import { validJSON } from "../helpers/helperFunctions";
+import { validateUser, validJSON } from "../helpers/helperFunctions";
 import { Recipe } from "./classes";
-import { Request, Response } from "express";
+import { Response } from "express";
 import { recipeModel } from "./schemas";
 
 export default class recipeCRUD {
 
-    async GET(id: number, res: Response): Promise<void> {        
-        const response = await recipeModel.findOne({ id: id });
-        res.send(response);
+    async GET(id: string, res: Response) {
+        const response = await recipeModel.findById(id);
+        if (response == null) return res.status(404).send("Object not found");
+        res.status(200).send(response);
+    }
+
+    async GET_ALL(res: Response) {
+        const response = await recipeModel.find();
+        res.status(200).send(response);
     }
 
     async POST(obj: any, res: Response) {
         if (validJSON(obj)) {
             try {
+                obj.addedBy = res.locals.user;
                 const classObj = new Recipe(obj);
                 const model = new recipeModel(classObj);
+                model.addedBy.user = res.locals.user;
                 await model.save();
-                res.sendStatus(200);
+                res.status(200).send("Recipe saved successfully");
             } catch (err) {
-                res.send(err);
+                res.status(400).send("Something went wrong");
             }
-        } else res.send("not a valid JSON");
-        //TODO update with correct error code
+        } else res.status(400).send("not a valid JSON");
     }
-    async PUT(id:number,obj:any, res: Response) {
-        //TODO: Guards against invalid input or ID
-        const before = await recipeModel.find({ id: id });
-        await recipeModel.findByIdAndUpdate(before, obj);
-        const test = await recipeModel.findOne({ id: id });
-        res.sendStatus(200);
+    async PUT(id: string, obj: any, res: Response) {
+        const before: Recipe | null = await recipeModel.findOne({ id: id });
+        console.log(before);
+
+        if (before == null) return res.status(404).send("Object not found");
+        if (validateUser(before, res.locals)) {
+            await recipeModel.findByIdAndUpdate(before, obj);
+            return res.status(200).send("Object succesfully changed");
+        } else {
+            return res
+                .status(403)
+                .send("You are not allowed to update this recipe");
+        }
     }
-    async DELETE(id:number,res:Response) {
-        //TODO: add guards to avoid deleting wrong obj
-        //TODO: add response when no obj is found
-        await recipeModel.deleteOne({ id: id });
-        res.sendStatus(200);
+    async DELETE(id: string, res: Response) {
+        const obj: any = recipeModel.findById(id);
+        if (obj == null) return res.status(404).send("Object not found");
+        if (validateUser(obj, res.locals)) {
+            await recipeModel.deleteOne({ id: id });
+            return res.status(200).send("Object succesfully deleted");
+        } else
+            return res
+                .status(403)
+                .send("You are not allowed to delete this recipe");
     }
-    
 }
